Don't copy local image if `qemu-img` available
If the source image is local, instead of copying it over, use qemu-img
command to create an image with the source image as backing image instead.
If `qemu-img` command is not available or the command fails for any
reason, we revert back to copying the image.

For crc, this reduces disk usage greatly (~ 9GB currently) and start up
time by around 20 seconds on my very high-end system with an SSD.

This adds an optional runtime dependency on `qemu-img` command.

Fixes https://github.com/code-ready/machine-driver-libvirt/issues/37