Remove DiskPath, use ImageFormat instead
DiskPath is currently provided by code using machine-driver-libvirt, but
it must match the path machine-driver-libvirt copied DiskPathURL to.

This is quite fragile, and it would be better if the path/name the VM
image was copied to was internal to machine-driver-libvirt.
This is what this commit achieves by making use of the newly added
'VMDriver.ImageFormat' to build the filename.