Add storage pool support
Until now, the libvirt machine driver directly creates an image file,
and then passes this path to libvirt.
This commit adds a libvirt storage pool where this file will live, this
will enable a later commit to implement something similar to `virsh
vol-resize` in order to be able to resize disk images.
This does not change anything related to how the files used by the
libivirt machine driver are laid out, it only 'registers' the
'~/.crc/machines/crc' directory with libvirt so that it knows this
location contains disk images (see 'virsh pool-list' and 'virsh
vol-list')