storage: Stop using incorrect StoragePoolDefineXML flag
The code is calling conn.StoragePoolDefineXML() with
libvirt.STORAGE_POOL_CREATE_NORMAL as a flag. This is incorrect as this
flag is meant for calls to conn.StoragePoolCreateXML(), not DefineXML().

libvirt.STORAGE_POOL_CREATE_NORMAL's value is 0 though, so this should
not be causing any problems. This commit removes its use anyway, and use
0 instead.