Update go modules to latest versions
github.com/libvirt/libvirt-go is untouched as we want to be 100% sure
that we don't link to symbols from libvirt.so which are only available
in new libvirt versions.
If we are not careful with this, the machine driver will fail to start on older
RHEL releases because it misses these too-new symbols.