Switch to using centos stream images from almalinux base-8
almalinux base-8 images are based on RHEL-8 and doesn't have go-1.23.x
which is required now to build the binary. Switching to centos stream9
image make sure latest version of golang available to consume. Removing
curl because it is part of base image

In downstream we already have go-1.23 attached to our target tag which
make sure it build downstream also.